Make redshift-gtk read output of the redshift child process.
This allows redshift-gtk to show the current state of the redshift
process. redshift-gtk follows the enable state of redshift and toggles
the icon accordingly.

The implementation is changed to use glib to spawn the child process
instead of relying on python subprocess module. This is necessary
because of inflexibility in the python module.