Add systemd .service file for redshift
Add a systemd user .service file to allow users to run redshift as a
daemon using systemd. Also adds a new configure option,
`--with-systemduserunitdir`, which gives the install path.

If set to `no`, systemd support is disabled.
If set to `yes`, we use `pkg-config` to find the path.

By default, the option is disabled, because if a user prefix is given,
but no other options are given, then the build may fail trying to
install to the system directory given by `pkg-config`.