Use a new protocol to ask the compositor authorization to change the gamma
A compositor may want to restrict the access to the gamma_control_manager
interface, to avoid clients changing the gamma without the user knowing.
With the new protocol extension redshift request permission to the
compositor to bind the gamma control global, so that the compositor
can grant/deny it, or ask the user for input.