Added backlight support via sysfs
Created a backlight_state struct to hold the configuration used to
change the backlight brightness via sysfs.
The backlight number of color_setting_t is a number between 0 and 1 to
control the monitor's backlight regardless of the absolute values used
by a specific driver.
This is still a prototype as there is not support to control de minimum
and maximum values, the sysfs path (controller) nor anything else.