Add a delay between restarts in systemd service
Sometimes, there's no "display-manager" service, the Xorg server
is then initialized after the TTY login (maybe with xorg-xinit).

This change gives some time (~30 seconds) to start the Xorg server
and starts redshift smoothly.

The defaults are:
    DefaultRestartSec=100ms
    DefaultStartLimitIntervalSec=10s
    DefaultStartLimitBurst=5

And if a service is more than 5 times within 10 seconds, it
will not be permitted to start any more.

Incrementing the `RestartSec` to 1s, and raising the `StartLimitIntervalSec`
to 30s, and `StartLimitBurst` to 30, it should give the user enough time to
login and start X.