geoclue2: ensure fallback to manual mode if unavailable
Even if support for GeoClue2 was included at compile time, it could be
missing at runtime. Be sure to check that the service is really
installed.

Tested on Arch Linux without GeoClue2 installed (which results in a
fallback to manual mode) and with GeoClue2 installed (which results in
automatically starting GeoClue2 and using it).