Status icon: Prefer symbolic icons when available
Use 'redshift-status-on-symbolic' and 'redshift-statys-off-symbolic'
if available.

If these icons are not present in the icon theme, fallback to
'redshift-status-on' and 'redshift-status-off'.

Symbolic icons are the only way for themes to properly support
both light and dark panels with grey icons. If the icon name ends
with -symbolic, GTK renders the foreground color of the icon based
on the color of the panel behind it.