Set redshift to run after graphical-sesion.target
The previous reference (`display-manager.service`) is not something
that's included with systemd. While it may be present on some user's
system (because they created it, or some specific package includes it),
it can't be expected to be present.

On the other hand, `graphical-session.target` is included with systemd,
and can be expected to be there, and be triggered when starting a
graphical session.

This guarantees that redshift will only run AFTER a graphic session is
up (eg: either Xorg or Wayland are up).