Change services to by wanted by graphical-session
The previous value (`default.target`) meant that redshift got started
when a user session began, at which point we can't expect a graphic
session to be up and running (eg: Xorg or Wayland may have not even been
invoked yet).

Setting `WantedBy=graphical-session.target` ensures that, when the
service is enable, it'll only get run when starting a graphical session
(again, either Xorg or Wayland, whatever DE's may be used).